docs: update example
